Force approval when getting Drive access
Without forcing an approval prompt, Google OAuth will only return a
refresh token the first time that you connect an application. If, for
whatever reason, you reconnect the application, Google will do a fast
no-prompt auth and won't return a refresh token, which is unfortunate
since that's the only credential we actually use.

By forcing the approval prompt, we ensure that we always get a refresh
token.