Make relativeTimeFormat tests pass in browser too
When running tests interactively in a browser (e.g. opening your
personal Chrome instance, not just using the HeadlessChrome from
`meteortesting:mocha`), if your locale observes DST, one test in
particular (the last of the 'formats correctly with multiple items'
suite) fails because the duration is different by one hour (the delta is
one hour longer).  In the full-browser environment, `process.env.TZ`
doesn't exist, and modifying it does not change the locale.  Instead, we
can just use `Date.UTC` to make the test explicitly UTC-oriented, which
avoids the off-by-one-hour issue in this unit test.

Also, this test suite was not actually correctly named -- probably a
copy-paste error from when this was first created.  So fix that while
we're here.