Enable @typescript-eslint typed linters
...and then turn several of them back off. A lot of the linters are
intended to tamp down use of `any`, which is admirable, but also
difficult to avoid in most of the places where we use it. (Although I
did go through and clean up some of the places where we didn't need to
rely on `any` first)

Most of the substantive changes were to account for the
`no-misused-promises` and `no-floating-promises` rules. This is because
in some cases we would invoke a promise as a way to fire off a
background task, without waiting for the result. For those cases, I
mostly used `void` casts, sometimes combined with IIFEs where necessary.