Keep previous `answers` value on Puzzles update
c3ffbee79cc53ea2eb184a8f99c76e0cc4ffa4af seems to have introduced a
regression in which every subsequent update to a `Puzzle` that does not
specify an explicit value for `answers` winds up causing `answers` to
be set to `[]`, stripping the denormalized answer values, which leads
PuzzleListPage to render it as unsolved (since PuzzleListPage relies on
the denormed `answers` rather than `Guess`es.

The real purpose of this `autoValue` implementation is to uppercase all
the answers.  We do not rely on it to backfill `answers: []` in the
event a caller did not provide that key.  To that end, we should return
`undefined` rather than `[]` from our `autoValue` implementation.

Fixes #943.