Remove gitRevision method (and associated machinery)
This was originally introduced for debugging purposes, but Meteor
includes a built-in primitive accomplishing the same goal
(`Meteor.gitCommitHash`). There's a small semantic difference between
the two (our revision always returned the running code on the server,
while `Meteor.gitCommitHash` returns the code on the client), but given
how Meteor's autorefresh works in production, they should always be the
same.