Unify the argument format of two relative time format methods
Switch `relativeTimeFormat` from accepting `complete` (a boolean) to
`maxElements` (an integer). In addition to bringing it into parity with
`terseRelativeTimeFormat`, it is also strictly more expressive (since
complete: false is equivalent to maxElements: 1, and complete: true is
equivalent to maxElements: -1).

However, the defaults remain different - the terse formatter defaults to
"complete" formats, while the non-terse formatter does not.