Adjust group sorting
Adds another couple criteria for more fine-grained group sort order to
address situations where two distinct metas have `meta-for:` the same
group (but no other obvious grouping or nesting relationship).  We also
take this opportunity to make sure that a group with all puzzles solved
is always punted to the end -- there's nothing left to be done with it.

Fixes #697.