Fix `updatePuzzle` callers
Without this patch, attempting to update a puzzle from the puzzle edit
modal (either on the PuzzlePage or the PuzzleListPage) results in a 400
error.

These callsites were incorrectly passing the additional argument
`huntId`, which is included in `PuzzleModalFormSubmitPayload` but not
accepted by the `updatePuzzle` method.

It is not clear to me why the typechecker didn't complain about passing
an incompatible type here, but maybe that's because in a structural
types world the extra key is permitted.