Use ValidateShape for TypedMethod calls
Since we generally require exact object matching when validating
TypedMethod arguments, we don't want to accidentally accept extra fields
in the argument type. By using ValidateShape, we can catch this at the
type level.

This also dramatically simplifies the implementation of `ValidateShape`,
and in the process improves the error reporting by marking the specific
fields that are superfluous as type `never`, rather than the entire
parameter.