Add className mb-3 to all FormGroups
Per the migration guide, you are now supposed to handle spacing with
margin utilities (the `m${pos}-${amount}` classes).  The standard margin
quantity (1x spacer) is 3, and we want margin on the bottom, so we now
get to shower all our forms with "mb-3".

Alternatively, we could make a class that adds the spacings, but I think
that'll just be worse in terms of render time, component hierarchy size,
and code size, so this seemed good enough.