Wrap link buttons with LinkContainer
The type declarations appear to indicate that the constraints on
Button's `as` argument reject use of react-router-dom's Link component.

We can work around this by using react-router-bootstrap's LinkContainer,
which hooks the onclick event to call react-router's navigate() (just as
the Link element would have), and just a standard intrinsic <a> element
for the button (the LinkContainer even propagates an appropriate href
value for us).

I suspect one of our dependencies (probably react-bootstrap) has an
incorrect type declaration, on account of the fact that the
implementation itself still worked when I passed `<Button as={Link}>`,
but that can be a problem for another day.