Preserve previous underline behavior
In bootstrap 4, links had no decoration by default, but added an
underline on hover.  This is generally nice and has a clean appearance
on desktop-class devices.

This default was changed in bootstrap 5, presumably to make links more
obvious when working on touch-based interfaces (where underline-on-hover
is not really an option).  However, I'm inclined to avoid changing our
UI's behavior at the same time as taking the library upgrade.  If we
decide we want to make which things are clickable links more obvious on
touch-controlled devices, we can do that in a separate PR.