Fix an incorrect missing dep on PuzzlePage
Before this change, it was possible that if you hit `F5` to reload
the page while you had a puzzle page open, you might hit a race
condition where:

* `activePuzzle` was still `undefined` despite the various subscriptions
  finishing loading, because
* `activePuzzle` did not declare a dep on `puzzleDataLoading` and the
  cached value (computed *before* the sub data had arrived) was
  initially undefined, and
* `useTracker` had not yet scheduled a rerun of that particular
  invalidated computation, because
* the subscription completion that had triggered the rerender of
  PuzzlePage was not the puzzle data subscription, but a different sub

Once hit, this would cause an exception in rendering where we'd attempt
to access `activePuzzle.title`, which was made possible in part by a
faulty non-null type assertion.

On top of all that, `activePuzzle` could still be undefined once
subscriptions all loaded if the URL contained an incorrect `puzzleId`,
which can happen if the user makes an error in copying a link.

This change should ensure that `activePuzzle` gets updated at the
right time, and as a bonus makes the rest of the page more
resilient to incorrect URLs.