Clean up RTCDebugPage for StrictMode
All of `<AccordionHeader>` is actually a button, so it's technically
forbidden to put a second button inside the header.

We had some buttons for copying unique IDs on the RTCDebugPage which did
exactly that.  This commit replaces them with a readonly input that
selects its contents on click and which prevents further event
propagation to avoid toggling the accordion when the target was the
input element.  The other `<ClipButton>`s in the `<AccordionBody>` are
kept as-is.

I also opted to remove the carets from the accordion headers, as:

* they were broken and did not actually update when the accordion
  changed collapse state, probably a bug introduced during the
  react-bootstrap 2 accordion rewrite, and
* they seem redundant now anyway with the react-bootstrap 2 Accordion
  design with the `<AccordionButton>` on the right hand side

Since the accordion component itself now also clearly displays which
element is active with a blue hue for the entire header, we have no need
to reimplement that ourselves with AccordionContext, so removing that
simplifies things a bit too.