Use ref/action form of OverlayTrigger as needed
OverlayTrigger needs to know the actual DOM node to attach to, and
passes a ref to its child.  If the thing you provide as a child isn't
a DOM node, and you don't forward refs, then OverlayTrigger falls back
to using findDOMNode, which isn't StrictMode-compliant.

This can be neatly resolved as described in
https://react-bootstrap.github.io/components/overlays/#customizing-trigger-behavior
by using the function form of OverlayTrigger.

The only place where we use OverlayTrigger with a child that does not
render as a DOM node is in NotificationCenter, where we have an
intermediate `<CopyToClipboard>` so that's what this commit fixes.