Allow overriding more env vars in run script
I have a VM that has too much RAM to not get opted into auto-clustering,
but I also run mongo on it and it has too few resources to handle the
write load of a bunch of backends, so I'd like to be able to explicitly
opt out of auto-clustering via setting an environment variable on the
docker instance.

I decided it was probably appropriate to allow a server operator to
override the `HTTP_FORWARDED_COUNT` for their deployment too, so that's
here as well.