Define our own Id, Email, and Url validation
For a couple of reasons, `simpl-schema` upstream has removed their copy
of these regexes in the latest version, which we'd like to be able to
upgrade to.  To unblock taking that upgrade, instead of relying on
`SimpleSchema.RegEx`, define our own equivalent regexes (for Id and
Email) and a custom validator for Url that just attempts to construct a
URL object (since that regex was vulnerable to regex DoS attacks).

Unblocks #1050.