Upgrade to TypeScript 4.8.4
This requires two fixes to account for TypeScript 4.8 being stricter
about generic functions. If you take a generic parameter and pass it to
a generic function, your generic needs to be as strict as the
declaration of the function.

For our `dropIndex` function, this is a semi-easy fix, although it's
hard to get at the `Document` type in the upstream mongodb library
without declaring an otherwise unnecessary dependency, so we just
copy-paste it here (it's fairly simple).

For `@restart/ui`, we need to manually override `react-bootstrap`'s
dependency for the time being, which is done by manually editing
`package-lock.json` and then letting `npm` clean it up. (With NPM 8,
this is natively supported, but Meteor isn't there yet.)