Migrate from launch configurations to launch templates
Launch templates are AWS's newer way of specifying an instance and are
replacing launch configurations. AWS recently sent out a notice that new
instance types won't be supported in launch configurations going
forward, so it's time for us to switch. The migration is fairly
straightforward, although it does require an extra layer of indentation
(thanks, YAML) which makes the diff look worse than it is. It's pretty
minimal with a word diff.