Take @types/react upgrade
There are a few changes in the v18 release of @types/react that we have
to account for. They've redefined `useCallback` to prevent inferring
`any` types, which we apparently were allowing in a few places - most
notably when we grab the `onChange` type from `react-bootstrap`'s
`FormControlProps` since that's nullable. They also deprecate
`ReactChild` and recommend replacing it with `ReactNode` instead.