Post a chat message on metas when a linked puzzle is solved
When a puzzle is solved, if that puzzle has any `group:` tags, and any
puzzles have the corresponding `meta-for:` tags, send a chat message to
those puzzles to let meta solvers know that the puzzle has been solved.

This requires a small tweak to the hook configuration to include the
newly correct answer as an argument for hooks to consume.

Fixes #592.