Implement the backend logic for remote mute
Remote mute is primarily handled by a new property on Peer -
`remoteMutedBy`. When non-null (set to the user ID of the user that
performed the remote mute), the client is prevented from updating
server-side mute state until it explicitly acknowledges the mute (at
which point the client is transitioned into normal muted state).

When a second client requests that someone be remote muted, we (a)
create an audit record of the remote mute, (b) set the `remoteMutedBy`
property on the Peer (and also set `muted` so that it gets displayed
correctly), and (c) find all producers associated with this Peer and
pause them.

When a client discovers that it has been remote muted, it updates its
local state to be muted and acknowledges the change in state. After
which, it can unmute itself. This is represented in a new `useCallState`
dispatch action, which means that in the future we can potentially track
more state (e.g. to display a modal informing the user that they have
been muted).