Fix bug in useCallState around server reconnection
This was most frequently hit in development, when modifying only server
code.  The server would stop and disconnect, and the client would
reconnect with the same hunt/call/tab, but because the producer track
from the captured media stream had not changed, even after the client
got updated transports, we would not start the new producer.

Attaching the transport ID associated with a given producer allows us to
notice when we need to recreate the transport.  Doing so requires that
we avoid mediasoup automatically closing the underlying track.  Since
we already correctly close the track ourselves when leaving a call, we
can simply pass `stopTracks = false`.