Enable "noUncheckedIndexedAccess" in TS config
By default, when you index into an array in TypeScript, it returns the
type of an element of that array. However, indexing out of bounds will
return `undefined`. Enabling "noUncheckedIndexedAccess" causes TS to
reflect that reality.

This requires fixing up a lot of places where we were indexing into
arrays. Ideally, I reworded the code so that the typechecker could tell
what was happening (e.g. replacing for loops with forEach loops), but in
other places where the logic made it clear, I just made use of the `!`
operator.