Build moira from source
Debathena seems to be sluggish about updating to newer Ubuntu releases
than 18.04, and with it about to go EOL, we need an alternative
strategy. While it would be great if we didn't have a dependency on
moira and blanche in the long term, this moves us forward in the short
term.