Fix up Dockerfile
It turns out that the behavior of Dockerfile heredocs with an explicitly
specified interpreter (i.e. `RUN <<EOF bash`) is a bit wonky and hard to
understand. Much clearer is putting a shebang line in the heredoc -
docker will handle that by spitting out a temporary file and executing
it; much clearer.

However, that doesn't seem to work with whitespace stripping, so just
unindent everything instead.

And finally, for consistency, switch all of the scripts to use bash
(instead of the few cases where we were using dash) and make sure to set
pipefail everywhere - even if it's not necessary, it's a good
belt-and-suspenders protection.