Require explicit guess direction/confidence
Range inputs don't support having an "unset" state, but we can simulate
one with a boolean field that we set to true when someone changes the
value. With this information, we can ensure that someone explicitly
chooses a direction and confidence before submitting a guess, which is
useful both in real time for operators evaluating guesses and for the
post-game analysis.

Fixes #301.