Track activity on Google Drive documents
This uses the Google Drive API's `files.watch` API method to receive
webhooks when changes are made to a document. We setup a watch on every
document in the database and handle renewing it as necessary.

When we're notified by Google that a document has been changed, we store
a record in the database - effectively a tuple of `(timestamp,
document)`. Absence of a record indicates lack of activity.

Google's notifications aren't especially timely - in practice they seem
to send notifications at about 3 minute granularity - so we use 5 minute
granularity for our own tracking. It's a nice round number, and pretty
well approximates the granularity of the notifications.

Activity records are currently never garbage collected. This should be
safe because generally puzzles don't see constant activity over the
course of a hunt - e.g. in 2022, puzzles were open for an average of 5
hours. Assuming perfect density, that's a total of roughly 12,000
records - not nothing, but not completely unmanageable. We can always
revisit in the future and add a expiration index if needed.

Since we're now asking an external service to reach out to us, we need
some way of addressing the server from the internet. In production, this
is easy. In development, if Meteor otherwise only knows itself to be
accessible over localhost, we attempt to integrate with ngrok to setup a
tunnel.

This commit only wires up collection of activity data; serving that to
clients and displaying it in a useful way is left as an exercise for the
reader (or more likely, me).