Rework guess queue page
The primary motivation for reworking this page was to disallow changing
guesses to most puzzle states. Instead of allowing any state, we now
only allow non-pending guesses to be changed back to pending.

On its own, that would make the page quite sparse, so it's a good
opportunity to rethink how we display the firehose of guesses from the
ground up. Going forward, at large screen sizes, we now display them in
a table (using CSS grid), with one line per guess (instead of 5 lines
per guess). At small screen sizes, we revert to a multi-line display
that's more similar to the old formatting (still using CSS grid, but in
a much more limited capacity).

In both cases, the display is significantly more visual: we use the
puzzle piece and skull and crossbones icons for links to puzzle and JR
pages (respectively). We also switch to using CopyToClipboard (and a
copy icon) instead of having an input that highlights itself on focus
for you to manually copy from. The two slider inputs (solve direction
and confidence) are shown here with sliders instead of purely
numerically. And we introduce icons for the various guess states.