Enrich guesses with "intermediate" state and additional notes
Oftentimes, a puzzle will spit out an intermediate answer - maybe an
instruction or a cluephrase. These intermediate answers can sometimes be
submitted, resulting in either confirmation or an additional step to
take. Historically, we've marked such instructions as "rejected", which
isn't exactly true. This lets us track them as correct, just not as a
final solution.

This also adds an "additionalNotes" field on guesses, which can be used
to capture whatever reply or additional instructions we get. It can also
be used for rejected answers to capture why the operator rejected it.

Additional notes (and the new intermediate state) are correctly
displayed when present, but there is no logic for specifying them yet.
However, this is enough for us to capture some of the intermediate
answers for our fixture hunt.