Do earlier exit in gdriveDocumentWatcher if feature flagged off
If the feature flag is disabled, we can avoid attempting to import ngrok
at all, which prevents a noisy error around unhandled promise rejection
when running without optional dependencies.

Additionally, we can put the bit where we attempt to resolve ngrok in a
try block so we issue a warning rather than have a top-level promise throw
if we've avoided installing optional dependencies.