Make guess notifications linger briefly after processing
This involves two pieces: one on the backend publication, and one on the
frontend in NotificationCenter.

In the backend, we add a "linger time" option to JoinPublisher, where
when items disappear from the initial query, rather than immediately
propagate the refcount decrement, we wait for the linger timeout to
pass.  In the case for `pendingGuesses`, this means the client will
observe the guess transition state, and then later it (and potentially
the other records is references) will be removed from the publication.

In the frontend, we track both `state: "pending"` and recently-updated
`Guess`es, and display the recently-updated records for 4 seconds after
their `updatedAt` time, disable the action buttons, but saturate the
button indicating the state into which the guess was placed.

Fixes #186.