Fix incorrect `observeChanges` usage in JoinedObjectObserver
This code appears to have expected the `fields` given by the
`observeChanges` callbacks to behave as `observe` does, where all field
values are provided, but they are not.

This caused spurious decref calls on the foreign-keyed values when the
top-level object changed but did not change the foreign keys.  `fields`
would not contain any fields associated with foreign keys, and we would
not incref them, but did decref them.

We should only update the foreign-keyed fields on change() when the
field is present in the `fields` object.  Meteor's documentation
specifies that if a field is removed from the object, the key will be
present in `fields` with value `undefined`.  So we should check if the
`fields` object has the property to determine if we need to take action
for this key, and then depending on the value we can act appropriately.