Eliminate "random_by" feature flags
These would mostly be useful if we ever wanted to do A/B testing or some
sort of partial rollout, but we've never wanted that and it's hard to
imagine wanting it in the future. Given that our only current use of
feature flags is as circuit-breakers, we can make the logic much
simpler.