Switch from watching individual files to fetching all changes
Google Drive imposes secret rate limits on setting individual file
watches, to the point that it's difficult to set up watches on all files
in our production instance (roughly 1500 files) without running afoul of
them.

So instead, we switch to using the `changes.list` method, which
enumerates changes to all files. I initially overlooked this due to a
poor understanding of its behavior, but in practice it has all of the
data we need.

We could use `changes.watch` to get notified when new changes are
available, but those notifications aren't sent out very frequently, and
Google Drive's rate limits are high enough that we can just poll roughly
every 15 seconds without introducing significant latency on updates. We
use a single lock to ensure that only one server is polling for changes
at a time, to ensure that multiple servers don't inadvertently stack up
against the rate limits.

Therefore, for now, we no longer need any of the incoming webhook
infrastructure from the previous implementation. (If we pursue a
publicly available Jolly Roger instance in the future such that users
are required to link their own Google Drives for document storage, we
may need to fetch changes from enough distinct drives that we need to
revisit this.)