Avoid use of the same <Tooltip> `id` in multiple places
Each of the following ids are used by both the `GuessQueuePage` data
table as well as the `NotificationCenter` toast:

- `guess-${guess._id}-direction-tooltip`
- `guess-${guess._id}-confidence-tooltip`
- `guess-${guess._id}-copy-tooltip`

We should avoid reusing ids on the page.  One way to help achieve that
is to encode something about the file in which the tooltip appears in
the id string itself, so I switched the ones in the NotificationCenter
to include `notification-`.

These were the only reused `id`s I found.