Add huntActivity publish for client
This more or less completely reworks the activity tracking from what was
added in the previous branch (#1236). This leaves us with a
significantly simpler database schema, which we can then use to build
the in-memory aggregation required to usefully present this data to the
client.

Instead of capturing `RecentActivity` and `ConsolidatedActivity`
records, we now only capture `CallActivity` records (which serves the
same function as `RecentActivity`, but is specialized to only call data;
there is no pre-aggregation of chat data). This has the upside of making
all of our primary sources for activity (`CallActivity`,
`DocumentActivity`, and `ChatMessage`) functionally immutable. Then,
instead of aggregating these records into `ConsolidatedActivity`, we
perform the aggregation in memory.

We ensure that only one aggregator exists for a given hunt, but allow
multiple client subscriptions to be attached to it. The aggregator
shuffles the data first by puzzle and subsequently by bucket, keeping
track of the unique users that we've observed in each bucket. It also
automatically schedules timeouts for when a bucket of data falls off of
our expiration window, and removes it from its cache. Then it publishes
the data in a pseudo-collection with the unique user sets projected down
to counts.