Add an in-house lint rule to block sync mongo methods
This introduces (but does not yet enable) a custom embedded ESLint rule
that detects calls to the Mongo API that use the legacy synchronous
interfaces. It requires a custom ESLint rule (as opposed to something
like eslint-plugin-ban) because some of the method names are suitably
generic that we can't reliably detect them without type information.

Most of the work in the rule is determining whether we are actually
calling methods on a Mongo collection or cursor, since we have to walk
the inheritance tree. None of these APIs are well-documented, but this
logic runs on our codebase without erroring out, so that'll have to be
good enough.

There's not really any way to add an ESLint plugin without putting it
into a separate, require-able package, which requires a little
gymnastics on our part. We can write our rules in TypeScript (and
consistent with the overall project style), but only if our package has
a compilation step (both for TypeScript compilation, and because ESLint
`require`s packages, which means they can't use import, so we need
module transpilation as well). We can pull in a local directory as a
dependency, but NPM doesn't run its build step or pull in any transitive
dependencies.

Therefore, we have to add any dependencies that the plugin needs to
Jolly Roger's overall dependencies, and we need to trigger the plugin's
build-step from Jolly Roger's overall `prepare` script (which runs after
`npm install`).