Configure no-extra-parens lint rule
For safety reasons, the autofixer for the new no-sync-mongo-methods rule
adds a bunch of potentially spurious parentheses. The no-extra-parens
rule does a good job of cleaning those parentheses up without changing
the semantics of the code.

Before enabling no-sync-mongo-methods, do one round of autofixing to
bring existing code in line with the no-extra-parens style.