Enable and autofix no-sync-mongo-methods
Other than the change to the ESLint config, the changes in this commit
consist entirely of running `eslint --fix` on the codebase.

This is not sufficient to pass linting on its own for a few reasons:
- The `forEach` method on `Mongo.Cursor` isn't auto-fixable
- This adds `await` without adding `async` to a bunch of methods

Committing this as a standalone change to make it easier to review the
rest.