Allow for-of loop syntax
The Airbnb style guide forbids using for-of loops. It warns that the
runtime overhead required for generators is high, and encourages using
forEach methods on arrays instead. However, in cases where we are
iterating over results from a database query (e.g. where we previously
would have used forEach), an async for-of loop is the only way to
iterate without first fetching the entire result set into memory. And
Meteor enables the regenerator runtime by default, so that's not
relevant either.

This enables for-of loops (and also makes a small correction to Meteor's
type declarations to allow for-of loops to work).