Make RemoteMuteConfirmModal visible on first render
In order to avoid the overhead of the remote mute modal until we
actually need it, we don't render it to the DOM until we're ready to
display it. If it defaults to being hidden when it is first rendered,
then you have to click on someone twice in order to mute them.