Fix update to DriveChangesPageTokens
This logic attempted to avoid updating the token record if another
process had gotten to it first, but this introduced two problems. First,
we didn't always use the token that we got back from the database.
Second, if another process _had_ gotten to it first, then our query
wouldn't match any existing record, and we would try to insert (which
would be a duplicate key error).