Ignore whitespace when looking for duplicate answers
We give a client-side warning when a user attempts to submit an answer
that has already been submitted. However, this warning only looks for
exact equality, meaning that we won't warn if the only difference is
whitespace.

Since this is only a warning and not a block, it seems acceptable if
this results in some false-positives, but it should have prevented
duplicate answers that were ultimately rejected in hunts past.