Introduce Google Apps Script integration
There are some functions that we'd like to take advantage of, aren't
exposed directly via Google's APIs, but _are_ exposed via Google App
Script - most notably, the ability to insert an image into a Google
Spreadsheet.

This commit introduces embedded source code for a Google Apps Script
project, along with logic to actually create and populate that project.
Apps Script projects can be exposed as a "web app", meaning they can
accept HTTP requests, so our project is structured as a JSON-RPC API. We
generate a shared secret for the project, and then use that secret to
authenticate requests to the API.

Unfortunately, it is not possible to do all of the configuration via the
API (in particular, it's not possible to create a Web App deployment),
so this still requires some amount of user involvement. A new section in
the setup page walks admins through the process.

Additionally, since we might make changes to the source code in the
future, we keep track of what source code is currently deployed to
Google, and present a notification if that is out of sync with our
embedded source.