Rework and share direction/confidence on guess queue
Instead of displaying the direction and confidence as sliders, show the
direction as an arrow + the numeric value, and show the confidence as a
progress bar. These are significantly more compact, and let us drop the
minimum width for the full-sized version of the page from large to
medium.

Additionally, pull these new components out so that they can be reused.