Add a bit of padding to puzzle activity UI
A bit between the door and the time.

A bit between the activity count and the sparkline.

The sparkline tooltip now has space on the bottom. I also made the
title fit on one line and the time legend is in smaller text.

One thing I'd like but haven't done is make the sparkline labels' baseline
match the base of the sparklines. TBD.