Improve our handling of empty strings
Historically, we've played fairly fast and loose with the distinction
between empty strings and undefined - both were false-y, so we'd often
treat both equivalently (e.g. by running them through the `||`
operator). This was often motivated by React, where having an input with
an undefined value is problematic (since React treats it as
uncontrolled), so we'd often cast undefined to an empty string.

Of course, they're not equivalent. This has popped up in various ways,
including switching more broadly to the nullish-coalescing operator in
4eab028a.

This commit generally looks for places where we cast undefined to empty
strings, and ensures that we cast those empty strings back to undefined
when we pass them to server code. There were also a few cases where the
server code didn't correctly handle false-y values (e.g. by using `$set`
instead of `$unset`); those are fixed as well.

Finally, we introduce a requirement that user's are no longer allowed to
set their displayName to an empty string, which would previously have
been allowed. We additionally enforce that it must not start with
whitespace. This is in anticipation of supporting @-mentions in the
future, where whitespace at the start of the display name would make it
difficult to auto-complete the name.