Use Drive Activity API for document activity
This replaces our old method of using the Drive list changes API to find
changes to instead using the Drive Activity API, which exposes higher
granularity data and lets us see who made an edit.

There are a few challenges to this API compared to the old
implementation. Records are returned newest-to-oldest, so there's no way
to ask for "records newer than what we've seen." Work around that by
tracking the most recent event timestamp we've seen, and specifying a
filter that the timestamp must be greater than that. (But include a 5
minute buffer just in case the API inserts records out of order.)

Since the API gives us user data, we add a new user field to
DocumentActivity (and rework the unique index to include it). The API
gives us back a series of Google account user IDs. However, there are
various reasons that we might not be able to map that to a Jolly Roger
user, so we allow the user field to be blank, indicating a user we
weren't able to identify.

Finally, we rework the published activity bucket to include a count of
document editors, instead of just a boolean.