More consistency in spacing and markup in notifications
Oddly, rejected answers were part of a Form, whereas intermediate
correct answers were just in a fragment. Neither used a proper form
label, which meant that the label, when clicked, would not focus the
input box.

This patchset deduplicates the markup for the two states that support
additional notes, and ensures both make correct semantic usage of
labels.

Also, I switched from `px` to the ~equivalent in `rem`, since that's the
unit that bootstrap uses for margin/padding.