Configure bugsnag server-side reporting
If the BUGSNAG_API_KEY environment variable is passed in (which we'll
attempt to set from credstash in production), then configure Bugsnag's
error reporting on server-side code.

By default, Bugsnag will report uncaught exceptions and unhandled
promise rejections, so we get some utility out of just this basic
configuration. However, once it's configured, we can also use the
Bugsnag library to report caught and handled exceptions.

Additionally, push the API key to the client for client-side error
reporting. (Note that the "notifier" API key is intended for both
client- and server-side reporting, and thus is not considered
sensitive.)